pipeline-od-vlm/pyproject.toml
[build-system]
requires = [
    "setuptools",
]
build-backend = "setuptools.build_meta"

[project]
requires-python = ">=3.10"
name = "pipeline_od_vlm"
readme = "README.md"
dynamic = [
    "version",
]
dependencies = [
    "ipython>=8.10",
    "jupyterlab>=3.0",
    "notebook",
    "kedro~=1.2.0",
    "kedro-mlflow>=2.0.2",
    "torch>=2.10.0",
    "datasets>=4.8.3",
    "opencv-python-headless>=4.13.0.92",
    "transformers>=5.3.0",
    "kedro-datasets>=9.2.0",
    "accelerate>=1.13.0",
    "kedro-viz>=12.3.0",
    "torchvision>=0.25.0",
]

[project.scripts]
pipeline-od-vlm = "pipeline_od_vlm.__main__:main"

[project.entry-points."kedro.hooks"]

[project.optional-dependencies]
dev = [
    "pytest-cov>=3,<7",
    "pytest-mock>=1.7.1, <2.0",
    "pytest~=7.2",
    "ruff~=0.12.0",
]

[tool.setuptools.dynamic.version]
attr = "pipeline_od_vlm.__version__"

[tool.setuptools.packages.find]
where = [
    "src",
]
namespaces = false

[tool.kedro]
package_name = "pipeline_od_vlm"
project_name = "pipeline-od-vlm"
kedro_init_version = "1.2.0"
tools = "['Linting']"
example_pipeline = "False"
source_dir = "src"

[tool.ruff]
line-length = 88
show-fixes = true

[tool.ruff.format]
docstring-code-format = true

[tool.ruff.lint]
select = [
    "F",
    "W",
    "E",
    "I",
    "UP",
    "PL",
    "T201",
]
ignore = [
    "E501",
]

[tool.kedro_telemetry]
project_id = "489216094bf34c929879391ad1d87c17"
